refactor to prompt api modal when setting options for strategy
